Add files via upload
The use of computer as a tool for painting is an emerging area of artistic exploration. This practice is generally referred as Digital Painting. In digital painting an artist or the user uses the computer I/O Devices like Mouse, keyboard, Monitors, Graphic Plotters or printers etc. to get his desired result. To be able to produce digital painting there is the need to have appropriate software such as MS- Paint, Adobe Photoshop or Coral Painter. These software provide users with several tools and options for getting variety of results. But as we know now is era of open source So, we wanted to develop an open source application which can provide users a simple functionality of these apps for free resulting which we are trying to develop a paint app in this project.